+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+0.3.0 (2019-12-15)
+------------------
+* Generate `id` property from `name` for every callback. The `id` is now used to
+  identify the callback, while `name` is used in the layout for displaying.
+
 0.2.0 (2019-11-03)
 ------------------
 * Use cookiecutter to create a proper project structure.